requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-nested-admin==4.1.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-monkey-business==1.1.0
sqlparse==0.5.3
tzdata==2025.1
gunicorn==21.2.0
supabase==2.15.2
scholarly==1.7.11
requests-html==0.10.0
fake-useragent==1.4.0
django-object-actions==5.0.0